refactor(userspace-dp): #6386 extract frag_assoc sibling
Move the NAT / NAT64 forward fragment-association install & consult
helpers out of poll_descriptor/mod.rs into a new pub(super) sibling
frag_assoc.rs: nat64_install_forward_fragment_assoc,
nat64_consult_forward_fragment_assoc, nat_install_forward_fragment_assoc,
nat_consult_forward_fragment_assoc, and the #6122 fail-closed same-family
discriminator flowless_fragment_requires_nat_translation.

Pure code motion, attr-verbatim: the four association helpers keep their
#[inline] and flowless_fragment_requires_nat_translation keeps its
deliberate #[cold] #[inline(never)]. Fn bodies byte-identical to their
prior location (verified with git diff --color-moved). No #[inline] is
added in this commit.

Cross-sibling edges are made explicit (the reject_reply.rs:17 pattern):
frag_assoc imports source_nat_would_translate_fragment from nat_exception
and prerouting_ingress_scope from prerouting_scope. The now-dead
source_nat_would_translate_fragment import is pruned from mod.rs (its
only caller moved). mod.rs re-imports the five helpers via
`use frag_assoc::{...}` so every loop call site is textually unchanged;
the stranded #3020 policy_packet_icmp doc reattaches to its fn.

Validation: cargo build (release) clean; moved bodies byte-identical.
